Run bathtub inundation model

Run bathtub inundation model

model_inundation(
  model,
  elev,
  elev_units,
  overlay = NULL,
  overlay_quantile = 0.95,
  from_elevation = -3,
  to_elevation = 3,
  step = 0.5,
  minimum_area = 0.6,
  minimum_area_units = "km^2",
  min_elev_cutoff = -5,
  use_outlets = T,
  model_ponding = T,
  site_name,
  workspace,
  overwrite = T
)

Arguments

model
A bathtub model. Result of assemble_network_model

elev
RasterLayer DEM object for inundation modeling. Result of DEM_setup.

elev_units
Units of elev values

overlay
sf object or RasterLayer of flooding extent (Optional)

overlay_quantile
Quantile of elevation values flooded by overlay to use for modeling flooding (Optional)

from_elevation
Lower bound of elevation for modeling. Units same as model inverts

to_elevation
Upper bound of elevation for modeling. Units same as model inverts

step
Step of sequence between from_elevation and to_elevation. Units same as model inverts

minimum_area
Minimum area of flooded area to keep

minimum_area_units
Units of flooded area

min_elev_cutoff
Minimum cutoff for elevation values

use_outlets
Force outlets to be connected by receiving waters (i.e., contiguous flooded area > min_elev_cutoff)

model_ponding
Model surface ponding from structure surcharge?

site_name
Name of site

workspace
Path to bathtub folder

Value

A list of sf objects denoting impacted infrastructure from each step of the model. Objects include impacted Pipes, Nodes, and Structures (using propagation through network), Nodes and Structures impacted by overland flooding with no propagation through network (e.g., "np_nodes" & "np_structures"), overland flooding, overland ponding. Includes the overlay if used rather than a range of water levels.

Examples

# bft_model_output <- model_inundation(
# model = bft_model,
# elev = bft_elev,
# elev_units = "m",
# from_elevation = -3,
# to_elevation = 4,
# step = 3/12,
# model_ponding = T,
# site_name = "beaufort",
# overwrite = T,
# minimum_area = 0.01,
# workspace = workspace
# )